requirements.txt
asgiref==3.7.2
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
Django==4.2.2
djangorestframework==3.14.0
drf-yasg==1.21.5
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
pip==23.1.2
pycparser==2.21
pyOpenSSL==23.0.0
PySocks==1.7.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
setuptools==67.8.0
simplejson==3.19.1
SQLAlchemy==2.0.17
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wheel==0.38.4
win-inet-pton==1.1.0
whitenoise==6.5.0
pandas==2.0.3
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-pool==3.1.7
python-dateutil==2.8.2
geopandas==0.13.2
pandas==2.0.3
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-pool==3.1.7
python-dateutil==2.8.2
geopandas==0.13.2
scikit-learn==1.3.0
xgboost==1.7.6
numpy==1.24.4
selenium==4.8.2
parameterized==0.9.0
webdriver_manager==4.0.0
pymcdm==1.1.0
extcolors==1.0.0
geojson==3.0.1
colormap==1.0.4
Pillow==10.0.0
easydev==0.12.1
django-cors-headers==4.1.0
